+++ src/components/combo_box.test.jsx
@@ -4543,56 +4543,6 @@ describe('renderOption', () => {
   });
 });
 
-describe('renderGroupAccessibleLabel', () => {
-  it('allows the group accessible label to be replaced', async () => {
-    render(
-      <ComboBoxWrapper
-        options={[{ label: 'foo', group: 'bar' }]}
-        renderGroupAccessibleLabel={(props) => (
-          <dl
-            data-foo="bar"
-            {...props}
-          />
-        )}
-      />,
-    );
-    await userEvent.tab();
-    expect(screen.getByRole('option').firstChild.tagName).toEqual('DL');
-    expect(screen.getByRole('option').firstChild).toHaveAttribute(
-      'data-foo',
-      'bar',
-    );
-  });
-
-  it('is called with context and props', async () => {
-    const spy = jest.fn(() => null);
-    render(
-      <ComboBoxWrapper
-        options={[{ label: 'foo', group: 'bar' }]}
-        renderGroupAccessibleLabel={spy}
-        test="foo"
-      />,
-    );
-    await userEvent.tab();
-    expect(spy).toHaveBeenLastCalledWith(
-      expect.objectContaining({
-        children: 'bar ',
-      }),
-      {
-        'aria-autocomplete': 'none',
-        'aria-busy': false,
-        expanded: true,
-        search: null,
-        notFound: false,
-        currentOption: null,
-        suggestedOption: null,
-        group: expect.objectContaining({ label: 'bar' }),
-      },
-      expect.objectContaining({ options: expect.any(Array), test: 'foo' }),
-    );
-  });
-});
-
 describe('renderValue', () => {
   it('allows the value to be replaced', async () => {
     render(
@@ -4816,19 +4766,6 @@ describe('renderNotFound', () => {
       },
       expect.objectContaining({ options: expect.any(Array), test: 'foo' }),
     );
-  });
-});
-
-describe('visuallyHiddenClassName', () => {
-  it('allows custom props', async () => {
-    render(
-      <ComboBoxWrapper
-        options={[{ label: 'foo', group: 'bar' }]}
-        visuallyHiddenClassName="bar"
-      />,
-    );
-    await userEvent.tab();
-    expect(screen.getByRole('option').firstChild).toHaveClass('bar');
   });
 });